+++ MiniApp/Classes/core/JavascriptBridge/MiniAppJavaScriptEnumerations.swift
@@ -35,6 +35,7 @@ public enum MiniAppCustomPermissionType: String, Codable, CaseIterable {
     case userName = "rakuten.miniapp.user.USER_NAME"
     case profilePhoto = "rakuten.miniapp.user.PROFILE_PHOTO"
     case contactsList = "rakuten.miniapp.user.CONTACT_LIST"
+    case accessToken = "rakuten.miniapp.user.ACCESS_TOKEN"
     case deviceLocation = "rakuten.miniapp.device.LOCATION"
 
     public var title: String {
@@ -45,6 +46,8 @@ public enum MiniAppCustomPermissionType: String, Codable, CaseIterable {
             return "Profile Photo"
         case .contactsList:
             return "Contact List"
+        case .accessToken:
+            return "Access Token"
         case .deviceLocation:
             return "Device Location"
         }