+++ Tests/Unit/MiniAppStatusTests.swift
@@ -68,13 +68,16 @@ class MiniAppStatusTests: QuickSpec {
                         permissionName: MiniAppCustomPermissionType(rawValue: MiniAppCustomPermissionType.profilePhoto.rawValue)!, isPermissionGranted: MiniAppCustomPermissionGrantedStatus.denied)
                     let contactListPermission = MASDKCustomPermissionModel(
                     permissionName: MiniAppCustomPermissionType(rawValue: MiniAppCustomPermissionType.contactsList.rawValue)!, isPermissionGranted: MiniAppCustomPermissionGrantedStatus.denied)
-                    miniAppKeyStore.storeCustomPermissions(permissions: [userNamePermission, profilePhotoPermission, contactListPermission], forMiniApp: "123")
+                    let accessTokenPermission = MASDKCustomPermissionModel(
+                    permissionName: MiniAppCustomPermissionType(rawValue: MiniAppCustomPermissionType.accessToken.rawValue)!, isPermissionGranted: MiniAppCustomPermissionGrantedStatus.denied)
+                    miniAppKeyStore.storeCustomPermissions(permissions: [userNamePermission, profilePhotoPermission, contactListPermission, accessTokenPermission], forMiniApp: "123")
                     var miniAppCustomPermissionList = miniAppKeyStore.getCustomPermissions(forMiniApp: "123")
                     expect(miniAppCustomPermissionList[0].permissionName.rawValue).to(equal("rakuten.miniapp.user.USER_NAME"))
                     expect(miniAppCustomPermissionList[0].isPermissionGranted.rawValue).to(equal("ALLOWED"))
                     expect(miniAppCustomPermissionList[1].permissionName.rawValue).to(equal("rakuten.miniapp.user.PROFILE_PHOTO"))
                     expect(miniAppCustomPermissionList[1].isPermissionGranted.rawValue).to(equal("DENIED"))
                     expect(miniAppCustomPermissionList[2].isPermissionGranted.rawValue).to(equal("DENIED"))
+                    expect(miniAppCustomPermissionList[3].permissionName.rawValue).to(equal("rakuten.miniapp.user.ACCESS_TOKEN"))
                     profilePhotoPermission.isPermissionGranted = .allowed
                     miniAppKeyStore.storeCustomPermissions(permissions: [profilePhotoPermission], forMiniApp: "123")
                     miniAppCustomPermissionList = miniAppKeyStore.getCustomPermissions(forMiniApp: "123")